+++ core/views.py
@@ -18,7 +18,8 @@ import directory_forms_api_client.helpers
 from directory_constants.choices import COUNTRY_CHOICES
 from directory_constants import urls
 from directory_components.helpers import get_user_country, SocialLinkBuilder
-from directory_components.mixins import CMSLanguageSwitcherMixin, GA360Mixin, CountryDisplayMixin
+from directory_components.mixins import (
+    CMSLanguageSwitcherMixin, GA360Mixin, CountryDisplayMixin, EnableTranslationsMixin)
 
 from core import forms, helpers, constants
 from core.context_modifiers import register_context_modifier, registry as context_modifier_registry
@@ -860,8 +861,8 @@ class PathRedirectView(QuerystringRedirectView):
         return f'{self.root_url}/{path}'
 
 
-class BusinessEnvironmentGuideFormView(GA360Mixin, InternationalHeaderMixin, FormView):
-    template_name = "core/business_environment_guide_form.html"
+class BusinessEnvironmentGuideFormView(EnableTranslationsMixin, GA360Mixin, InternationalHeaderMixin, FormView):
+    template_name = "core/investment_prospectus_form.html"
     form_class = forms.BusinessEnvironmentGuideForm
     subject = "Business Environment Guide Form"
     success_url = '/international/about-uk/why-choose-uk/business-environment-guide/success/'
@@ -911,7 +912,7 @@ class BusinessEnvironmentGuideFormView(GA360Mixin, InternationalHeaderMixin, For
 
 
 class BusinessEnvironmentGuideFormSuccessView(InternationalView):
-    template_name = 'core/business_environment_guide_form_success.html'
+    template_name = 'core/investment_prospectus_form_success.html'
     page_type = 'BusinessEnvironmentGuideFormSuccessPage'
     header_section = tier_one_nav_items.ABOUT_UK
     header_sub_section = tier_two_nav_items.WHY_CHOOSE_THE_UK
@@ -926,7 +927,7 @@ class BusinessEnvironmentGuideFormSuccessView(InternationalView):
         return super().dispatch(request, *args, **kwargs)
 
 
-class WhyBuyFromUKFormView(GA360Mixin, InternationalHeaderMixin, FormView):
+class WhyBuyFromUKFormView(GA360Mixin, EnableTranslationsMixin, InternationalHeaderMixin, FormView):
     template_name = "core/why_buy_from_the_uk_form.html"
     form_class = forms.WhyBuyFromUKForm
     subject = "How we help Guide Form"
@@ -1009,7 +1010,7 @@ def handler500(request, *args, **kwargs):
     return render(request, '500.html', status=500)
 
 
-class InternationalContactTriageView(GA360Mixin, InternationalHeaderMixin, FormView):
+class InternationalContactTriageView(GA360Mixin, EnableTranslationsMixin, InternationalHeaderMixin, FormView):
     template_name = 'core/contact_international_triage.html'
     form_class = forms.InternationalRoutingForm
     success_url = urls.domestic.CONTACT_US + 'international/'